Make UserImpl inner class of UserFactoryImpl
